service/requirements.txt
alibi-detect==0.8.0
bentoml==0.13.1
mlflow-skinny==1.21.0
numpy==1.19.2
pandas==1.2.4
scikit-learn==0.24.1
xgboost==1.4.2
